+++ main.go
@@ -6,6 +6,8 @@ import (
 	"net/http"
 
 	"github.com/anton-ag/todolist/internal/config"
+	"github.com/anton-ag/todolist/internal/database"
+	"github.com/anton-ag/todolist/internal/handlers"
 	"github.com/go-chi/chi/v5"
 
 	_ "modernc.org/sqlite"
@@ -15,28 +17,28 @@ func main() {
 	var config config.Config
 	config.Init()
 
-	initDb()
-	db, err := sql.Open("sqlite", "scheduler.db")
+	err := database.InitDB(config.DBFile)
 	if err != nil {
-		log.Fatal("Ошибка соединения с БД")
+		log.Fatalf("%v", err)
+	}
+	db, err := sql.Open("sqlite", config.DBFile)
+	if err != nil {
+		log.Fatalf("Ошибка соединения с БД: %v", err)
 	}
 	defer db.Close()
 
-	// router
 	r := chi.NewRouter()
 
 	fs := http.FileServer(http.Dir("web"))
 	r.Handle("/*", http.StripPrefix("/", fs))
+	r.Get("/api/nextdate", handlers.NextDate)
+	r.Post("/api/task", handlers.NewTask(db))
+	r.Get("/api/tasks", handlers.GetTasks(db))
+	r.Get("/api/task", handlers.GetTask(db))
+	r.Put("/api/task", handlers.UpdateTask(db))
+	r.Post("/api/task/done", handlers.DoneTask(db))
+	r.Delete("/api/task", handlers.DeleteTask(db))
 
-	r.Get("/api/nextdate", NextDateHandler)
-	r.Post("/api/task", NewTaskHandler(db))
-	r.Get("/api/tasks", GetTasksHandler(db))
-	r.Get("/api/task", GetTaskHandler(db))
-	r.Put("/api/task", PutTaskHandler(db))
-	r.Post("/api/task/done", PostDoneHandler(db))
-	r.Delete("/api/task", DeleteTaskHandler(db))
-
-	// launch server
 	err = http.ListenAndServe(config.Port, r)
 	if err != nil {
 		panic(err)
